+++ xmp/AsyncSemaphore.js
@@ -1,13 +1,9 @@
 #!/usr/bin/env node
 if ( require.main === module )
 {
-  var T      = require ( "../src/Tango" ) ;
-  var Event  = require ( "../src/Event" ) ;
-  var Client = require ( "../src/Client" ) ;
-  var Admin  = require ( "../src/Admin" ) ;
-  var Semaphore   = require ( "../src/Semaphore" ) ;
+  var gepard = require ( "gepard" ) ;
 
-  if ( T.getProperty ( "help" ) )
+  if ( gepard.getProperty ( "help" ) )
   {
     console.log (
       "Gepard example: Semaphore, acquire a given semaphore.\n"
@@ -17,7 +13,7 @@ if ( require.main === module )
     ) ;
     process.exit() ;
   }
-  new Admin().isRunning ( function admin_is_running ( state )
+  new gepard.Admin().isRunning ( function admin_is_running ( state )
   {
     if ( ! state )
     {
@@ -28,9 +24,9 @@ if ( require.main === module )
   });
   function execute()
   {
-    var key = T.getProperty ( "name", "user:4711" ) ;
-    var auto = T.getProperty ( "auto" ) ;
-    var sem = new Semaphore ( key ) ;
+    var key = gepard.getProperty ( "name", "user:4711" ) ;
+    var auto = gepard.getProperty ( "auto" ) ;
+    var sem = new gepard.Semaphore ( key ) ;
     console.log ( "Aquiring semaphor=" + key ) ;
     sem.acquire ( function ( err )
     {